move StatsQuery to a separate file
