+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/oracle/StatsTaskListGenerator.java
@@ -16,25 +16,21 @@
  */
 package com.google.edwmigration.dumper.application.dumper.connector.oracle;
 
-import com.google.auto.value.AutoValue;
 import com.google.common.collect.ImmutableList;
-import com.google.common.io.Resources;
 import com.google.edwmigration.dumper.application.dumper.ConnectorArguments;
 import com.google.edwmigration.dumper.application.dumper.task.Task;
 import java.io.IOException;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
 class StatsTaskListGenerator {
 
-  private static final ImmutableList<StatsQuery> QUERIES =
+  private static final ImmutableList<OracleStatsQuery> QUERIES =
       ImmutableList.of(
-          StatsQuery.create("hist-cmd-types", StatsSource.STATSPACK),
-          StatsQuery.create("app-schemas-pdbs", StatsSource.METADATA),
-          StatsQuery.create("app-schemas-summary", StatsSource.METADATA)
+          OracleStatsQuery.create("hist-cmd-types", StatsSource.STATSPACK),
+          OracleStatsQuery.create("app-schemas-pdbs", StatsSource.METADATA),
+          OracleStatsQuery.create("app-schemas-summary", StatsSource.METADATA)
           // TODO: add entries for other SQLs to this list
           );
 
@@ -42,7 +38,7 @@ class StatsTaskListGenerator {
   ImmutableList<Task<?>> createTasks(ConnectorArguments arguments) throws IOException {
     ImmutableList.Builder<Task<?>> builder = ImmutableList.<Task<?>>builder();
 
-    for (StatsQuery item : QUERIES) {
+    for (OracleStatsQuery item : QUERIES) {
       builder.add(StatsJdbcTask.fromQuery(item));
     }
     return builder.build();
@@ -58,25 +54,6 @@ class StatsTaskListGenerator {
 
     StatsSource(String value) {
       this.value = value;
-    }
-  }
-
-  @AutoValue
-  abstract static class StatsQuery {
-
-    abstract String name();
-
-    abstract StatsSource tool();
-
-    static StatsQuery create(String name, StatsSource tool) {
-      return new AutoValue_StatsTaskListGenerator_StatsQuery(name, tool);
-    }
-
-    @Nonnull
-    String queryText() throws IOException {
-      String path = String.format("oracle-stats/%s/%s.sql", tool().value, name());
-      URL queryUrl = Resources.getResource(path);
-      return Resources.toString(queryUrl, StandardCharsets.UTF_8);
     }
   }
 }
